fix: ImageSelector can now return null or undefined depending on which is needed
in json, undefined fields are just ignored. if you set a field
undefined, the field is just not included in the json.

in order to delete an image in backend, one needs to send imagefield:
null. also, in order to not change an imagefield while changin other
fields, setting it undefined if the value is just a url is preferred.
in patch requests, this field will not change. in put requests, if the
field is required, it may be undesirable still. but I think this is the
lesser evil way of solving this.

feat: update and create functions are actually added to detailview.